[Patchset 11] Remove warning suppressions and fix many casting bugs
This change removes most warning suppressions, in particular
from production code. In the past we inherit a bunch of -Wno-xxx
suppressions required to build gtest and libprotobuf headers.
Doing so, however, caused the suppressions to propagate back to
the translation units that were including any protobuf header
or any auto-generated .pb.h stub.
This change moves the gtest and probobuf header to be a
system include (-isystem vs -I). Doing so implicitly blacklists
any compiler warning on the headers.
This CL then re-enables warnings and deals with the fall out of
fixes that came out of this.

Bug: 77316877
Test: pefetto_unittests / perfetto_integrationtests

+++ include/perfetto/ipc/basic_types.h
@@ -31,8 +31,8 @@ namespace perfetto {
 namespace ipc {
 
 using ProtoMessage = ::google::protobuf::MessageLite;
-using ServiceID = uint64_t;
-using MethodID = uint64_t;
+using ServiceID = uint32_t;
+using MethodID = uint32_t;
 using ClientID = uint64_t;
 using RequestID = uint64_t;
